+++ openrem/remapp/admin.py
@@ -14,7 +14,7 @@ from remapp.models import (GeneralStudyModuleAttr,
     CtRadiationDose, CtAccumulatedDoseData,
     CtIrradiationEventData, ScanningLength,
     CtDoseCheckDetails, CtXRaySourceParameters,
-    Exports, SizeUpload)
+    Exports, SizeUpload, UniqueEquipmentNames, DicomStoreSCP, DicomRemoteQR)
 
 from django.contrib import admin
 
@@ -53,6 +53,9 @@ admin.site.register(CtDoseCheckDetails)
 admin.site.register(CtXRaySourceParameters)
 admin.site.register(Exports)
 admin.site.register(SizeUpload)
+admin.site.register(UniqueEquipmentNames)
+admin.site.register(DicomStoreSCP)
+admin.site.register(DicomRemoteQR)
 
 from django.contrib.auth.admin import UserAdmin
 from django.contrib.auth.models import User