+++ openrem/remapp/netdicom/qrscu.py
@@ -13,25 +13,26 @@ For help on usage,
 python qrscu.py -h 
 """
 
+import logging
 
 # call back
 def OnAssociateResponse(association):
-    print "Association response received"
+    logging.info("Association response received")
 
 
 def OnAssociateRequest(association):
-    print "Association resquested"
+    logging.info("Association resquested")
     return True
 
 def _move_req(my_ae, remote_ae, d):
     assocMove = my_ae.RequestAssociation(remote_ae)
-    print "Move association requested"
-    print "d is {0}".format(d)
+    logging.info("Move association requested")
+#    print "d is {0}".format(d)
     gen = assocMove.StudyRootMoveSOPClass.SCU(d, 'STOREOPENREM', 1)
     for gg in gen:
-        print "gg is {0}".format(gg)
+        logging.info("gg is %s", gg)
     assocMove.Release(0)
-    print "Move association released"
+    logging.info("Move association released")
 
 def _try_query_return(rsp, tag):
     try:
@@ -51,7 +52,7 @@ def _query_series(my_ae, remote_ae, d2, studyrsp):
     d2.Modality = ''
     d2.NumberOfSeriesRelatedInstances = ''
 
-    print 'd2: {0}'.format(d2)
+#    print 'd2: {0}'.format(d2)
 
     assoc_series = my_ae.RequestAssociation(remote_ae)
 
@@ -59,14 +60,14 @@ def _query_series(my_ae, remote_ae, d2, studyrsp):
 
     query_id = uuid.uuid4()
 
-    print 'In _query_series'
+#    print 'In _query_series'
     seRspNo = 0
 
     for series in st2:
         if not series[1]:
             continue
         seRspNo += 1
-        print "Series Response {0}: {1}".format(seRspNo, series[1])
+#        print "Series Response {0}: {1}".format(seRspNo, series[1])
         seriesrsp = DicomQRRspSeries.objects.create(dicom_qr_rsp_study=studyrsp)
         seriesrsp.query_id = query_id
         # Mandatory tags
@@ -196,7 +197,7 @@ def qrscu(
     query.save()
 
     # create association with remote AE
-    print "Request association with {0} {1} {2}".format(rh, rp, aec)
+#    print "Request association with {0} {1} {2}".format(rh, rp, aec)
     assoc = MyAE.RequestAssociation(RemoteAE)
 
     if not assoc:
@@ -206,14 +207,14 @@ def qrscu(
         query.save()
         MyAE.Quit()
         return
-    print "assoc is ... {0}".format(assoc)
+    logging.info("assoc is ... %s", assoc)
 
     # perform a DICOM ECHO
-    print "DICOM Echo ... ",
+    logging.info("DICOM Echo ... ")
     echo = assoc.VerificationSOPClass.SCU(1)
-    print 'done with status "%s"' % echo
+    logging.info('done with status %s', echo)
 
-    print "DICOM FindSCU ... ",
+    logging.info("DICOM FindSCU ... ")
     d = Dataset()
     d.QueryRetrieveLevel = "STUDY"
     d.PatientName = ''
@@ -328,12 +329,13 @@ def qrscu(
                         if s.series_description != 'Dose Info':
                             s.delete()
 
-    print "Release association"
+    logging.info("Release association")
     assoc.Release(0)
 
     # done
     MyAE.Quit()
     query.complete = True
+    query.stage = "Query complete"
     query.save()
 
 
@@ -423,72 +425,3 @@ def movescu(query_id):
     query.save()
 
     my_ae.Quit()
-
-
-# (0008, 0018) SOP Instance UID                    UI:
-# (0008, 0020) Study Date                          DA: ''
-# (0008, 0050) Accession Number                    SH: ''
-# (0008, 0052) Query/Retrieve Level                CS: 'SERIES'
-# (0008, 0060) Modality                            CS: ''
-# (0008, 0062) SOP Classes in Study                UI:
-# (0008, 0070) Manufacturer                        LO: ''
-# (0008, 1010) Station Name                        SH: ''
-# (0008, 1030) Study Description                   LO: ''
-# (0008, 103e) Series Description                  LO: ''
-# (0010, 0020) Patient ID                          LO: ''
-# (0020, 000d) Study Instance UID                  UI: 1.3.6.1.4.1.23849.1666612369.116.1635621988901632614
-# (0020, 000e) Series Instance UID                 UI:
-# (0020, 0011) Series Number                       IS: ''
-# (0020, 1202) Number of Patient Related Series    IS: ''
-# (0020, 1204) Number of Patient Related Instances IS: ''
-# (0020, 1206) Number of Study Related Series      IS: ''
-# (0020, 1208) Number of Study Related Instances   IS: ''
-# (0020, 1209) Number of Series Related Instances  IS: ''
-
-
-
-# If one or more than one modality selected:
-#     Query with first modality in 'ModalitiesInStudy' field
-#         1/ Respected as match, repeat with other modalities
-#         2/ Not respected as match, but returned in rsp - not likely
-#         3/ Ignored, returned all modalities
-
-# Detect 1/ because none of rsp have ModalitiesInStudy that don't include modality, or because field is not empty?
-# If 3/, don't repeat query with other modality selected - will get same results
-# If 3/, need to filter on rsp before presenting results or requesting move
-
-# Always go to series level query, and do move on series level.
-
-
-# From series level query function
-
-        # if series[1].Modality == 'SR':
-        #     # Not sure if they will be SR are series level but CT at study level?
-        #     # If they are, send C-Move request
-        #     #print "C-Move request for series with modality type SR"
-        #     if move:
-        #         _move_req(MyAE, RemoteAE, series[1])
-        #     continue
-        # seNum = str(series[1].SeriesNumber)
-        # if (seNum == '502') or (seNum == '998') or (seNum == '990') or (seNum == '9001'):
-        #     # Find Siemens (502 CT, 990 RF) and GE (998 CT) RDSR or Enhanced SR
-        #     # Added 9001 for Toshiba XA based on a sample of 1
-        #     # Then Send a C-Move request for that series
-        #     #print "C-Move request for series with number {0}".format(seNum)
-        #     if move:
-        #         _move_req(MyAE, RemoteAE, series[1])
-        #     continue
-        # try:
-        #     if series[1].SeriesDescription == 'Dose Info':
-        #         # Get Philips Dose Info series - not SR but enough information in the header
-        #         #print "C-Move request for series with description {0}".format(series[1].SeriesDescription)
-        #         if move:
-        #             _move_req(MyAE, RemoteAE, series[1])
-        #         continue
-        # except AttributeError:
-        #     # Try an image level find?
-        #     pass
-        # # Do something for Toshiba CT...
-
-# More rambling thoughts...
-#