+++ openrem/remapp/tools/get_values.py
@@ -27,6 +27,7 @@
 ..  moduleauthor:: Ed McDonagh
 
 """
+import logging
 
 def get_value_kw(tag,dataset):
     """Get DICOM value by keyword reference.
@@ -103,4 +104,8 @@ def get_or_create_cid(codevalue, codemeaning):
                 code_meaning = codemeaning,
                 )
             cid.save()
-        return ContextID.objects.get(code_value=codevalue)
+        code = ContextID.objects.filter(code_value__exact = codevalue)
+        if code.count() > 1:
+            logging.warning("Duplicate entry in the ContextID table: %s/%s, import continuing",
+                            codevalue, codemeaning)
+        return code[0]